Recipes and users refactoring

+++ backend/foodgram/users/views.py
@@ -2,13 +2,14 @@ from django.contrib.auth import get_user_model
 from djoser.views import UserViewSet
 from rest_framework import status
 from rest_framework.decorators import action
+from rest_framework.exceptions import NotAcceptable, NotFound
 from rest_framework.mixins import ListModelMixin
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 
 from .models import UserFollow
-from .serializers import UserFollowSerializer, UserSubscriptionSerializer
+from .serializers import UserSubscriptionSerializer
 
 User = get_user_model()
 
@@ -22,38 +23,39 @@ class CustomUserViewSet(UserViewSet):
         permission_classes=[IsAuthenticated],
     )
     def subscribe(self, request, id=None):
-        user_to_follow = self.get_object()
-        data = {
-            "follower": request.user.id,
-            "following": user_to_follow.id,
-        }
+        follower = request.user
+        following = self.get_object()
 
-        serializer = UserFollowSerializer(
-            data=data,
-            context=self.get_serializer_context(),
-        )
-        serializer.is_valid(raise_exception=True)
+        is_subscribed = UserFollow.objects.filter(
+            follower=follower,
+            following=following,
+        ).exists()
 
         if request.method == "GET":
-            UserFollow.objects.create(
-                follower=request.user,
-                following=user_to_follow,
-            )
+            if follower == following:
+                raise NotAcceptable("Нельзя подписаться на самого себя.")
+            if is_subscribed:
+                raise NotAcceptable("Такая подписка уже есть.")
+            UserFollow.objects.create(follower=follower, following=following)
             response_data = UserSubscriptionSerializer(
-                user_to_follow, context={"request": request}
+                following,
+                context={"request": request},
             )
             return Response(response_data.data)
 
         if request.method == "DELETE":
+            if not is_subscribed:
+                raise NotFound("Пользователь не подписан.")
+
             UserFollow.objects.filter(
-                follower=request.user,
-                following=user_to_follow,
+                follower=follower,
+                following=following,
             ).delete()
             return Response("OK", status.HTTP_201_CREATED)
 
 
 class SubscriptionViewSet(ListModelMixin, GenericViewSet):
-    queryset = User.extended_objects.with_recipes_count()
+    queryset = User.extended_objects.with_recipes_count().order_by("id")
     permission_classes = [IsAuthenticated]
     serializer_class = UserSubscriptionSerializer
